feat: Clear front-end layouts and structure.
