+++ routes/admin.php
@@ -1,0 +1,5 @@
+<?php
+
+use Mj\PocketCore\Router;
+
+Router::get('/admin', 'admin.index');
